+++ crates/ruma-common/src/profile.rs
@@ -13,9 +13,12 @@ use crate::SecondsSinceUnixEpoch;
 use crate::{OwnedMxcUri, PrivOwnedStr};
 
 mod profile_field_value_serde;
+mod static_profile_field;
+mod user_profile;
 
 #[doc(hidden)]
 pub use self::profile_field_value_serde::ProfileFieldValueVisitor;
+pub use self::{static_profile_field::*, user_profile::*};
 
 /// The possible fields of a user's [profile].
 ///